Create migrations table ReceiveMoneyList and update table ReceiveMoney, change method CQRS in module ReceiveMoney, change method update in module TransferMoney when update AttachmentList & TagList

+++ TabpediaFin/TabpediaFin/Domain/TransferMoney/TransferMoney.cs
@@ -4,7 +4,7 @@ public class TransferMoney : BaseEntity
 {
     public int TransferFromAccountId { get; set; } = 0;
     public int DepositToAccountId { get; set; }
-    public int Amount { get; set; } = 0;
+    public Int64 Amount { get; set; } = 0;
     public string Memo { get; set; } = string.Empty;
     public string TransactionNumber { get; set; } = string.Empty;
     public DateTime TransactionDate { get; set; }
